Add ComplexHypixelObject for property methods
